Move AppContextProvider interface to application package - Refs #25055
