+++ src/main/java/eionet/meta/controllers/AttributeFixedValuesController.java
@@ -1,5 +1,6 @@
 package eionet.meta.controllers;
 
+import eionet.meta.application.AppContextProvider;
 import eionet.util.CompoundDataObject;
 import eionet.meta.application.errors.DuplicateResourceException;
 import eionet.meta.application.errors.MalformedIdentifierException;
